04/09/24 The roles are working add, list, delete

+++ src/main/java/com/example/taskmanagement/model/Role.java
@@ -11,7 +11,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id_role;
+    private Long id_role;
 
    // @ManyToOne(fetch = FetchType.LAZY)
    // @JoinColumn(name = "idrole")
